changes book componet and home
